+++ src/components/loader.js
@@ -1,0 +1,18 @@
+// @ts-nocheck
+
+import React, { Fragment } from "react"
+
+import loadergif from "../../static/loading-opaque.gif"
+
+function Loader() {
+  return (
+    <Fragment>
+      <img
+        src={loadergif}
+        alt="..."
+        style={{ margin: "auto", display: "block", width: "60%" }}
+      />
+    </Fragment>
+  )
+}
+export default Loader
